% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovPCAscores.R
\name{aovPCAscores}
\alias{aovPCAscores}
\title{Plot ANOVA-PCA Scores from a Spectra Object}
\usage{
aovPCAscores(spectra, so, submat = 1, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra}}.}

\item{so}{\emph{List} of pca results created by \code{\link{aov_pcaSpectra}}.}

\item{submat}{Integer.  Selects list element \code{submat} from \code{PCA}
which is a list of PCA results, each corresponding to the computation
in \code{\link{aov_pcaSpectra}}.}

\item{\dots}{Additional parameters to be passed to \code{\link{plotScores}}.
For example, you can plot confidence ellipses this way.  Note that ellipses
are drawn based on the groups in \code{spectra$groups}, but the separation
done by \code{aov_pcaSpectra} is based on argument \code{fac}.  These may
not correspond, but you can edit \code{spectra$groups} to match if necessary.}
}
\value{
The returned value depends on the graphics option selected (see \code{\link{GraphicsOptions}}).
\describe{
 \item{base:}{A data frame or list containing the data plotted.  Assign the value and run \code{str()}
              or \code{names()} on it to see what it contains. Side effect is a plot.}
 \item{ggplot2:}{The plot is displayed, and a \code{ggplot2} plot object is returned if the
                 value is assigned. The plot can be modified in the usual \code{ggplot2} manner.
                 If you want the plotted values, you can access them via the base graphics mode.}
}
}
\description{
Uses the results from \code{\link{aov_pcaSpectra}} to plot the scores.
Argument \code{submat} is used to select PCA results from among those
stored in argument \code{PCA}.
}
\references{
Pinto, Bosc, Nocairi, Barros, and Rutledge. "Using ANOVA-PCA for
Discriminant Analysis: ..." Analytica Chimica Acta 629.1-2 (2008): 47-55.

Harrington, Vieira, Espinoza, Nien, Romero, and Yergey. "Analysis of
Variance--Principal Component Analysis: ..." Analytica Chimica Acta 544.1-2
(2005): 118-27.
}
\seealso{
The use of this function can be seen in
\code{\link{aov_pcaSpectra}}.  See also \code{\link{plotScores}}.
Additional documentation at \url{https://bryanhanson.github.io/ChemoSpec/}
}
\author{
Matthew J. Keinsley and Bryan A. Hanson, DePauw University.
}
\keyword{htest}
\keyword{multivariate}
