% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkGraphicsOpt.R
\name{chkGraphicsOpt}
\alias{chkGraphicsOpt}
\title{Checks the Graphic mode of the plots}
\usage{
chkGraphicsOpt(silent = TRUE)
}
\arguments{
\item{silent}{Silents the messages on change of graphics mode}
}
\description{
Checks the Graphic mode of the plots
}
