% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkReqGraphicsPkgsStub.R
\name{chkReqGraphicsPkgs}
\alias{chkReqGraphicsPkgs}
\title{Verify Required Graphics Packages are Available.}
\description{
This function is used by \code{ChemoSpec} and \code{ChemoSpec2D}, but is
formally part of \code{ChemoSpecUtils}.  You can access
full documentation via \code{\link[ChemoSpecUtils]{chkReqGraphicsPkgs}}.
}
