% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorSymbolStub.R
\name{colorSymbol}
\alias{colorSymbol}
\title{Color and Symbols in ChemoSpec and ChemoSpec2D}
\description{
This help page serves both \code{ChemoSpec} and \code{ChemoSpec2D}, but is
formally part of \code{ChemoSpecUtils}. You can access
full documentation via \code{\link[ChemoSpecUtils]{colorSymbol}}.
}
