% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conColSchemeStub.R
\name{conColScheme}
\alias{conColScheme}
\title{Change the Color Scheme of a Spectra or Spectra2D Object}
\description{
This help page serves both \code{ChemoSpec} and \code{ChemoSpec2D}, but is
formally part of \code{ChemoSpecUtils}. You can access
full documentation via \code{\link[ChemoSpecUtils]{conColScheme}}.
}
