% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkReqPkgs.R
\name{.chkReqPkgs}
\alias{.chkReqPkgs}
\title{Verify Required Packages are Available.}
\usage{
.chkReqPkgs(pkgs)
}
\arguments{
\item{pkgs}{Character. Vector of package names.}
}
\value{
Logical \code{TRUE}, invisibly, if all packages are locally available.
Otherwise, one or more messages about missing packages and \code{FALSE}.
}
\description{
\emph{Internal function}.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
