% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicsOutput.R
\docType{data}
\name{GraphicsOutput}
\alias{GraphicsOutput}
\title{Graphic options in ChemoSpec}
\description{
In \code{ChemoSpec}, the user may use any graphic mode from 'base' and 'ggplot2' for plotting.
The graphic mode can be found using \code{chkGrafOpt()}.
Initially the graphic option is set to 'base'.
The graphic options can be changed by doing \code{options(ChemoSpecGraphics = 'option')}, where 'option' is one of 'base' or 'ggplot2'
}
