\name{hmapSpectra}
\alias{hmapSpectra}

\title{
Create a Seriated Heat Map Comparing Samples and Spectral Data for a Spectra Object
}
\description{
Creates a heat map with marginal dendrograms using seriation procedures.  Very briefly, the samples that are most like each other occur in one corner, and the frequencies that are most informative with respect to the samples are in that corner as well. This is achieved by using heirchical cluster analysis and then re-ordrering the clusters in a coordinated way across each dimension.  See the reference.
}
\usage{
hmapSpectra(spectra, no.col = 5, title = "no title provided",
t.pos = c(0.5, 0.5, 0.5), cexRow = 1, cexCol = 1, ...)
}

\arguments{
  \item{spectra}{ An object of S3 class \code{"Spectra"}}.

  \item{no.col}{
The number of colors (levels) to use in the heat map.  Maximum of 9, generated by \code{RColorBrewer}.
}
  \item{title}{
A title for the plot.
}
  \item{t.pos}{
A vector of three numbers giving the x, y and adj for the title (adj controls justification; options are [0, 0.5, 1]).  x and y coordinates in [0...1] range.  Due to the way the underlying functions are written, you can't necessarily put the title just anywhere.  Play with it.
}
  \item{cexRow}{
Scale factor for the row labels.
}
  \item{cexCol}{
Scale factor for the column labels.
}
  \item{\dots}{
Additional arguments to be passed downstream.
}
}
\value{
None.  Side effect is a plot.
}
\references{
\url{http://academic.depauw.edu/~hanson/ChemoSpec/ChemoSpec.html}

}

\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}

\seealso{
\code{\link[seriation]{hmap}} which will get you to the package (there is no package index page); the vignette is a good place to begin (> vignette("seriation")).
}

\examples{
data(CuticleIR)
# trim it down so plot is not so crowded:
newIR <- removeFreq(CuticleIR, rem.freq =
CuticleIR$freq > 1500 | CuticleIR$freq < 500)
# remove some random samples too:
newIR <- removeSample(newIR, rem.sam = sample(1:157, 107))
# now make the heat map:
hmapSpectra(newIR, title = "Cuticle IR", t.pos = c(0.45, 0.25, 1),
cexRow = 0.5, cexCol = 0.1, no.col = 9)
}
\keyword{ multivariate }

