\name{metMUD1}
\alias{metMUD1}
\alias{metMUD2}
\docType{data}
\title{Made Up NMR Data Sets}
\description{
\code{metMUD1} is... \code{metMUD2} is ...
}
\usage{data(metMUD1)}
\format{
The data is stored as a \code{\link{Spectra}} object.
}
\source{
Created using various tools.  Contact the author if interested.}

\references{
\url{https://github.com/bryanhanson/ChemoSpec}
}

\examples{
data(metMUD1)
sumSpectra(metMUD1)
}
\keyword{datasets}
