% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScree2.R
\name{plotScree2}
\alias{plotScree2}
\title{Alternate Style Scree Plot DEPRECATED}
\usage{
plotScree2(pca, ...)
}
\arguments{
\item{pca}{\itemize{
  \item An object of class \code{\link{prcomp}}, modified to include a
list element called \code{$method}, a character string describing the
pre-processing carried out and the type of PCA performed (it appears on the
plot).  This is automatically provided if \code{ChemoSpec} functions
\code{\link[ChemoSpec]{c_pcaSpectra}} or \code{\link[ChemoSpec]{r_pcaSpectra}}
were used to create \code{pca}.
}}

\item{\dots}{Additional parameters to be passed to plotting functions.}
}
\value{
None.  Side effect is a plot.
}
\description{
This function is deprecated.  Please use \code{plotScree(pca, style = 'alt')}
}
