% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGroupsStub.R
\name{updateGroups}
\alias{updateGroups}
\title{Update Group Names in a Spectra or Spectra2D Object}
\description{
This function is used by \code{ChemoSpec} and \code{ChemoSpec2D}, but is
formally part of \code{ChemoSpecUtils}.  You can access
full documentation via \code{\link[ChemoSpecUtils]{updateGroups}}.
}
