% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check4GapsStub.R
\name{check4Gaps}
\alias{check4Gaps}
\title{Check for Discontinuities (Gaps) in a Vector & Optionally Make a Plot}
\description{
This function is used by \code{ChemoSpec} and \code{ChemoSpec2D}, but is
formally part of \code{ChemoSpecUtils}.  You can access
full documentation via \code{\link[ChemoSpecUtils]{check4Gaps}}.
}
