\name{ChemoSpec-package}
\Rdversion{1.1}
\alias{ChemoSpec-package}
\alias{ChemoSpec}
\docType{package}

\title{
Exploratory Chemometrics for Spectroscopy
}

\description{
A collection of functions for top-down exploratory data analysis of spectral data obtained via nuclear magnetic resonance (NMR), infrared (IR) or Raman spectroscopy.  Includes functions for plotting and inspecting spectra, peak alignment, hierarchical cluster analysis (HCA), principal components analysis (PCA), model-based clustering and statistical total correlation spectroscopy (STOCSY) analysis.  Robust methods appropriate for this type of high-dimensional data are available.  ChemoSpec is designed with metabolomics data sets in mind, where the samples fall into groups such as treatment and control.  Graphical output is formatted consistently for publication quality plots.  ChemoSpec is intended to be very user friendly and help you get usable results quickly.  A vignette covering typical operations is available.

}

\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec}
}
\keyword{ package }
\keyword{ multivariate }
