% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemoSpec-package.R
\docType{package}
\name{ChemoSpec-package}
\alias{ChemoSpec-package}
\alias{ChemoSpec}
\title{Exploratory Chemometrics for Spectroscopy}
\description{
A collection of functions for top-down exploratory data analysis of spectral
data obtained via nuclear magnetic resonance (NMR), infrared (IR) or Raman
spectroscopy.  Includes functions for plotting and inspecting spectra, peak
alignment, hierarchical cluster analysis (HCA), principal components
analysis (PCA) and model-based clustering. Robust methods appropriate for
this type of high-dimensional data are available.  ChemoSpec is designed
with metabolomics data sets in mind, where the samples fall into groups such
as treatment and control.  Graphical output is formatted consistently for
publication quality plots.  ChemoSpec is intended to be very user friendly
and help you get usable results quickly.  A vignette covering typical
operations is available.
}
\seealso{
Additional documentation at \url{https://bryanhanson.github.io/ChemoSpec/}
}
\author{
Bryan A. Hanson (DePauw University), Tejasvi Gupta & Matthew J. Keinsley.

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{multivariate}
\keyword{package}
