\name{ChemoSpec-package}
\Rdversion{1.1}
\alias{ChemoSpec-package}
\alias{ChemoSpec}
\docType{package}

\title{
Exploratory Chemometrics for Spectroscopy
}

\description{
 A collection of functions for plotting spectra (NMR, IR, Raman) and carrying out various forms of top-down exploratory data analysis, such as HCA, PCA, model-based clustering and STOCSY analysis.  Robust methods appropriate for this type of high-dimensional data are available.  ChemoSpec is designed with metabolomics data sets in mind, where the samples fall into groups such as treatment and control.  Graphical output is formatted consistently for publication quality plots.  ChemoSpec is intended to be very user friendly and help you get usable results quickly.  A vignette illustrating typical operations is available.
}

\details{
\tabular{ll}{
Package: \tab ChemoSpec\cr
Type: \tab Package\cr
Version: \tab 4.0.1\cr
Date: \tab 2015-03-16\cr
License: \tab GPL-3\cr
}
}

\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec}
}
\keyword{ package }
\keyword{ multivariate }
