\name{ChemoSpec-package}
\Rdversion{1.1}
\alias{ChemoSpec-package}
\alias{ChemoSpec}
\docType{package}

\title{
Exploratory Chemometrics for Spectroscopy
}

\description{
A collection of functions for plotting spectra (NMR, IR etc) and carrying out various forms of top-down exploratory data analysis, such as HCA, PCA and model-based clustering.  The design permits comparison of data from samples which fall into groups such as treatment vs. control.  Robust methods appropriate for this type of high-dimensional data are available.  ChemoSpec is designed to be very user friendly and suitable for people with limited background in R.  A vignette illustrating typical operations is available.
}

\details{
\tabular{ll}{
Package: \tab ChemoSpec\cr
Type: \tab Package\cr
Version: \tab 1.61-3\cr
Date: \tab 2013-09-20\cr
License: \tab GPL-3\cr
}
}

\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\references{
\url{http://academic.depauw.edu/~hanson/ChemoSpec/ChemoSpec.html}
}
\keyword{ package }
\keyword{ multivariate }
