% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkGraphicsOptStub.R
\name{chkGraphicsOpt}
\alias{chkGraphicsOpt}
\title{Check the Graphics Output Option/Mode}
\description{
This function is used by \code{ChemoSpec} and \code{ChemoSpec2D}, but is
formally part of \code{ChemoSpecUtils}.  You can access
full documentation via \code{\link[ChemoSpecUtils]{chkGraphicsOpt}}.
}
