\name{colLeaf}
\alias{colLeaf}
\title{
Color the Leaves of a Dendrogram
}
\description{
This function colors the leaves of a dendrogram object.  The code was taken from the help files.  Not for end-user use.
}
\usage{
colLeaf(n, spectra)
}
\arguments{
  \item{n}{
A node in a dendrogram object.
}
  \item{spectra}{
An object of S3 class \code{"Spectra"}.
}
}
\value{
Returns a node with the label color properties set.
}

\references{
\url{http://academic.depauw.edu/~hanson/ChemoSpec/ChemoSpec.html}
}

\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\keyword{ utilities }
\keyword{ cluster }
