% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorSymbol.R
\name{colorSymbol}
\alias{colorSymbol}
\title{Colors and Symbols in ChemoSpec and Spectra Objects}
\description{
In \code{ChemoSpec}, the user may use any color name/format known to R.  For
ease of comparison, it would be nice to plan ahead and use the same color
scheme for all your plots.  The current color scheme of a
\code{\link{Spectra}} object may be determined using \code{\link{sumGroups}}
or changed using \code{\link{conColScheme}}. Also,
\code{\link{splitSpectraGroups}} has another means of changing the color
scheme, but this is intended for the situations when you are creating new
categories/groups for your samples.
}
\seealso{
Additional documentation at \url{https://bryanhanson.github.io/ChemoSpec/}
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{color}
\keyword{utilities}
