\name{colorSymbol}
\Rdversion{1.1}
\alias{colorSymbol}
\title{
Colors and Symbols in ChemoSpec and Spectra Objects
}
\description{
In \code{ChemoSpec}, the user may use any color name/format known to R.  For ease of comparison, it would be nice to plan ahead and use the same color scheme for all your plots.  The current color scheme of a \code{"Spectra"} object may be determined using \code{\link{sumGroups}} or changed using \code{\link{conColScheme}}. Also, \code{\link{splitSpectraGroups}} has another means of changing the color scheme but this is intended for the situations when you are creating new categories for your samples.
}

\references{
\url{http://academic.depauw.edu/~hanson/ChemoSpec/ChemoSpec.html}
}

\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}

\keyword{ utilities }
\keyword{ color }
