% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoopThruSpectra.R
\name{loopThruSpectra}
\alias{loopThruSpectra}
\title{Display the Spectra in a Spectra Object One at a Time DEPRECATED}
\usage{
loopThruSpectra(spectra, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra}}.}

\item{\dots}{Parameters to be passed downstream.}
}
\value{
The returned value depends on the graphics option selected (see \code{\link{GraphicsOptions}}).
\itemize{
 \item{base:}{    None.  Side effect is a plot.}
 \item{ggplot2:}{    Returns a \code{ggplot2} plot object. The plot can be modified in the usual
                     \code{ggplot2} manner.}
 }
}
\description{
DEPRECATED: Use \code{reviewAllSpectra} instead.  Plots each spectrum in a
\code{\link{Spectra}} object one at a time, and
waits for a return in the console before plotting the next spectrum.  Use
\code{ESC} to get out of the loop.
}
\details{
The ggplot2 mode plots all the \code{\link{Spectra}} objects in the dataset on a single column.
}
\seealso{
See \code{\link{GraphicsOptions}}
         for more information about the graphics options. Additional documentation at
         \url{https://bryanhanson.github.io/ChemoSpec/}
}
\author{
Bryan A. Hanson (DePauw University)
}
\keyword{hplot}
