% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoresStub.R
\name{plotScores}
\alias{plotScores}
\title{Plot Scores from PCA, MIA or PARAFAC Analysis of a Spectra or Spectra2D Object}
\description{
This function is used by \code{ChemoSpec} and \code{ChemoSpec2D}, but is
formally part of \code{ChemoSpecUtils}.  You can access
full documentation via \code{\link[ChemoSpecUtils]{plotScores}}.
}
