% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScoresCor.R
\name{plotScoresCor}
\alias{plotScoresCor}
\title{Compute Confidence Ellipses}
\usage{
plotScoresCor(x, quan = 1/2, alpha = 0.025)
}
\arguments{
\item{x}{As per \code{\link[mvoutlier]{cor.plot}}.}

\item{quan}{As per \code{\link[mvoutlier]{cor.plot}}.}

\item{alpha}{As per \code{\link[mvoutlier]{cor.plot}}.}
}
\value{
A list with the following elements (a simpler version of that in the
original function \code{\link[mvoutlier]{cor.plot}}): \item{x.cls}{The x
values for the classical ellipse.} \item{y.cls}{The y values for the
classical ellipse.} \item{c}{The correlation value for the classical
ellipse.} \item{x.rob}{The x values for the robust ellipse.}
\item{y.rob}{The y values for the robust ellipse.} \item{r}{The correlation
value for the robust ellipse.}
}
\description{
A utility function which when given a x,y data set computes both classical
and robust confidence ellipses. An internal function, not generally called
by the user.
}
\author{
Bryan A. Hanson, DePauw University. Derived from \code{\link[mvoutlier]{cor.plot}}.
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec}
}
\seealso{
See function \code{\link[mvoutlier]{cor.plot}} in package
\pkg{mvoutlier} on which this function is based.
}
\keyword{multivariate}

