% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_pcaSpectra.R
\name{r_pcaSpectra}
\alias{r_pcaSpectra}
\title{Robust PCA of a Spectra Object}
\usage{
r_pcaSpectra(spectra, choice = "noscale")
}
\arguments{
\item{spectra}{An object of S3 class \code{\link{Spectra}}.}

\item{choice}{A character vector describing the type of scaling to be
carried out.  One of \code{c("noscale", "mad")}.}
}
\value{
An object of classes \code{converted_from_princomp} and \code{prcomp}.
It includes a list element called \code{$method}, a
character string describing the pre-processing carried out and the type of
PCA performed (used to annotate plots).
}
\description{
A wrapper which carries out robust PCA analysis on a \code{\link{Spectra}}
object.  The data are row- and column-centered, and the user can select
various options for scaling.
}
\examples{

\dontrun{
# This example assumes the graphics output is set to ggplot2 (see ?GraphicsOptions).
library("ggplot2")
data(metMUD1)
pca <- r_pcaSpectra(metMUD1)

p1 <- plotScree(pca)
p1

p2 <- plotScores(metMUD1, pca, pcs = c(1, 2), ellipse = "cls", tol = 0.05)
p2 <- p2 + ggtitle("Scores: metMUD1 NMR Data")
p2

p3 <- plotLoadings(metMUD1, pca, loads = 1:2, ref = 1)
p3 <- p3 + ggtitle("Loadings: metMUD1 NMR Data")
p3
}
}
\references{
K. Varmuza and P. Filzmoser \emph{Introduction to Multivariate
Statistical Analysis in Chemometrics}, CRC Press, 2009.
}
\seealso{
\code{\link[pcaPP]{PCAgrid}} for the underlying function,
\code{\link{c_pcaSpectra}} for classical PCA calculations,
\code{\link{s_pcaSpectra}} for sparse PCA calculations,
\code{\link{irlba_pcaSpectra}} for PCA via the IRLBA algorithm.
Additional documentation at \url{https://bryanhanson.github.io/ChemoSpec/}

\url{https://bryanhanson.github.io/ChemoSpec/}
}
\author{
Bryan A. Hanson (DePauw University)
}
\keyword{multivariate}
\keyword{robust}
