% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowDistStub.R
\name{rowDist}
\alias{rowDist}
\title{Compute Distance Between Rows of a Matrix}
\description{
This function is used by \code{ChemoSpec} and \code{ChemoSpec2D}, but is
formally part of \code{ChemoSpecUtils}.  You can access
full documentation via \code{\link[ChemoSpecUtils]{rowDist}}.
}
